Incur export eligibility
Working up https://github.com/p1m-ortho/xs-sebzer/issues/8.

Matters involved are also of importance
with respect to https://github.com/p1m-ortho/xs-sebzer/issues/1.

Whether or not any single page is susceptible
to this script’s parsing is obviously dependent
on the respecting page’s DOM. So far, I have
only made it so that the search query result pages
were parsed correctly. Fortunately, publication
collections («подборки публикаций») are built
with a similar DOM, thus being also successfully
processed with the current script.

However, any other page not sufficiently
similar in its DOM to the search results page,
single-record pages specifically so, will not be
adequately processed. Some of them will from
now on display a respecting warning on load—
some will not.

Implementation of single-record page parsing
is, as I see it, at risk of IP bans because of multiple
simultaneous page requests, so will abstain from it
for now, seeing that this functionality is available
with collection export support.